+++ crates/fluvio-cli/src/lib.rs
@@ -138,10 +138,11 @@ mod root {
         Version(VersionOpt),
 
         /// Generate command-line completions for Fluvio
-        #[structopt(
-            name = "completions",
-            settings = &[AppSettings::Hidden]
-        )]
+        /// {n}{n}Run the following two commands to enable fluvio command completions.
+        /// Open a new terminal for the changes to take effect.
+        /// {n}{n}$ fluvio completions bash > ~/fluvio_completions.sh
+        /// {n}$ echo "source ~/fluvio_completions.sh" >> ~/.bashrc
+        #[structopt(name = "completions")]
         Completions(CompletionCmd),
 
         /// Generate metadata for Fluvio base CLI